Simplify RunContext
